Set notification pid to pid of policy event (#613)
Currently, the pid of the notification event that denotes the start of a
sysdig capture is the agent pid. Change this to the pid of the
associated policy event when the capture is triggered by a security
policy event. This is done by adding the pid to the capture job request
and filling it in in security_mgr::start_capture().

Otherwise, set the notification pid to 0. This ensures that the
notification event is not filtered out by the default agent
functionality that skips activity for the agent's session id.