Events before captures (#626)
* Ensure policy events are sent before capture msgs

Make sure policy events (that have a capture action) are always sent
before the contents of the capture. We do this by adding a flag
m_defer_send that tells the capture job to not send any data until a
later JOB_SEND_START message is received.

To handle this split, the dump_job_request message, which was being used
for both start/stop requests, is now split into dump_job_request and an
embedded start_job_details object. It contains all the params like
before/after duration, filter, etc. It's only filled in when starting a
capture job.

In the security manager, after all actions are complete, and if one of
the actions was a capture action, send the JOB_SEND_START message.

In capture_job, m_defer_send is required to send any capture chunks. a
JOB_SEND_START message sets m_defer_send to false. A job will only be
removed once all chunks are sent, so they will wait for a SEND_START
message if needed.

Also, we need to make sure that the policy event message is always sent
to the collector before the first capture chunk message. A big cause for
this was that the policy events messages weren't using the same shared
queues as the other message types. Merge the policy events messages so
they are sent using the same blocking_queue object as all other
messages. This eliminates the synchronized_policy_events object from
everywhere.

Now, policy events messages are queued to the blocking queue with MEDIUM
priority. Since this is higher than the LOW priority used for the
capture chunks, and since MEDIUM messages are always sent before LOW
messages, they will be sent first.

One downside of this is that groups of policy events messages won't be
merged into a single message. However, we'll assume that policy event
throttling and the batching that occurs within the security_mgr code
will take care of the worst of this.

Additionally, properly handle throttled policy events + captures:

- If a policy event is throttled, all of the details on the event are
  lost, and instead a simple count by policy name is incremented. If the
  original policy had a capture action, we need to make sure to stop the
  associated capture job so it won't stick around.

- To address this, change accept_policy_event to return a bool that
  indicates whether the policy event was accepted. If it was *not*
  accepted, and if it had a capture action, stop the associated capture
  job.

Finally, a few changes for unit tests:

- Revive the security_messages unit test. This required adding necessary
  host_scope/container scope to the sample policies, simplifying the
  test a bit to wait exactly once for exactly 1 metrics message (or
  fail) at startup, and reduce the sample policies to only match the
  specific action used by the test.

- When starting a capture job, if m_send_initial_keepalive is true, send
  a keepalive message right away. Currently only used by tests as a way
  to coordinate activity and capture jobs.

One minor change: in security policy metrics, make sure metrics are
initially initialized. Ensures initial logged counts are 0.

* Add unit test for delayed start, many cleanups

Add a unit test for delayed capture starts. The test starts a capture
setting defer_send to true, waits 10 seconds, sends a send_start
message, and then waits for the entire capture to be sent.

Also make a lot of simplifications to the memdump unit tests in an
attempt to make them more reliable:

 - Instead of using the agent's sinsp_worker thread, which also includes
   a full analyzer + accompanying modules, use a test_sinsp_worker which
   only has an inspector + capture_job_handler event hook. This reduces
   the complexity of the test a lot.
 - As a consequence, many tests can be simpler about waiting for
   results, etc. which also reduces the time they wait for success or
   failure. e.g. queue_fetch() is removed, replaced by directly reading
   from the capture chunk queue. wait_dump_complete now times out based
   on time and not message count.
 - some other unused options such as limited capture sizes, etc. are
   removed.
 - After starting a capture, request an initial keepalive and wait for it
   before performing any activity. This makes sure that the file open,
   etc. actions that should occur after a job has started don't occur
   until the capture has started.
 - At setup, the sinsp_worker/capture_job_handler threads can race
   with the main test thread. To address this, wait for
   test_sinsp_worker::run to finish one loop before completing setup.

I also had to make changes to verify_no_dragent_events. With changes to
how events are written to the file in conjunction with skipping events
other than the test program itself, the trace created by
verify_no_dragent_events actually had no events in it, as the process
filter was proc.name=tests but none of those events were written to the
file.

For that test only, allow all events from all programs to be written to
the file. The test still verifies that no events from the test program
are actually in the file.