Fix concurrency issues in kubecollect
Change the kubecollect API to present a single external call
and move the reset of the logic inside the package. Return a
channel from whic the caller can read (and forward) congroup
messages.

Use the rpc context to cancel informers if needed, and wait for
them to finish before closing the events channel.

Partial: rename funcs so they will no longer be exported