Temporary solution for pinning gRPC version
Currently we have two directories on GOPATH build and src
We automatically generate *.pb.go files for draiosproto and sdc_internal
and put them on build path.

These files to compile need gRPC as dep and we cannot use a pinned version under vendor
because golang compiler does not accept that a file outside source directory has a dep
under vendor/ path. The dep needs to stay in the same folder of gopath.

This temporary solution bundles the *.pb.go files under cointerface src and adds the gRPC
version to the vendor/ folder.