Create a watchdog thread for detecting k8s api server errors
The SharedInformers we use will log errors when they can't connect to
the k8s api server, but there's no way to get a notification in code.
Directly doing a watch allows us to catch errors and kill the informers.

It's still possible to race and hit a few errors logs from informers,
but there should be at most 1 per informer. We should fix this but we'll
need to propose a change to the SharedInformer interface.