Send events for init containers
Init containers are treated the same as regular containers, and the logic for
all containers has changed. ADDED events are sent the first time we see a
container go to running. REMOVED events are sent for running to terminated
transitions (UpdateFunc) and for all containers on pod deletion. We may send
two REMOVED events for the same container, but that's safe with infra state.