Allow open->openat rewrite. SMAGENT-744
Recent Ubuntu releases (at least 17.10) use the openat() syscall
for libc open() calls. Adapt the test code to cope with it.

This commit fixes the sys_call_test.fs_openat test. On Ubuntu
the code calls openat() more times than expected so we need to
find the right calls instead of assuming that the openat() call
found is the one called explicitly in the test.