Allow open->openat rewrite. SMAGENT-744
Recent Ubuntu releases (at least 17.10) use the openat() syscall
for libc open() calls. Adapt the test code to cope with it.

This commit fixes the memdump tests to use a filter that will
catch both open< and openat> events (the name in openat is reported
in the enter event, unlike open).