Use mesos framework name as is instead of adding the uri
The previous reason to do that was probably to distinguish two different marathon
installation under the same cluster

The tradeoff is that when you have marathon leader election, the uri changes,
causing containers be assigned to another grouping name. Since it seems way more reasonable
to have just `marathon` as of now (like DC/OS deploys), I'm doing this change.