Use cpuacct cgroups for more accurate CPU time reporting. SMAGENT-690 (#680)
* Use cpuacct cgroups for more accurate CPU time reporting. SMAGENT-690

* Mount cpuacct cgroup subsystem in docker image. SMAGENT-690

SMAGENT-690 includes support for getting CPU usage from the cpuacct
cgroup. So when running the agent in a container, we need to mount
the cpuacct cgroup subsystem as well.

* Make mounting cgroups more robust. SMAGENT-690

If a cgroup controller is mounted together with a different one
(e.g. cpu,cpuacct), they cannot be mounted separately.

Parse /proc/self/mountinfo to find the cgroup subsys combinations
and use these to mount the same combinations in $SYSDIG_HOST_ROOT