Fix Java 9 version parsing. SMAGENT-464
With Java 9, the java.version property no longer looks like 1.x.y
but may simply be "9" (e.g. "9-Ubuntu"). Fix the parser to handle
this, i.e. chop off the string to parse as a double at the second
dot (like before) or the first dash (whichever comes earlier).

This makes the Java agent at least start with JDK 9 installed.
Getting it to work is a different thing.