Catch NoClassDefFoundError when patching LinuxVirtualMachine class. SMAGENT-464
Looks like this class is no longer present in Java 9. Catching this
exception makes JMX work for the following scenarios:

* Java 7/8 used in agent, Java 7/8/9 in monitored app, remote JMX enabled

* Java 7/8 used in agent, Java 7/8 in monitored app, remote JMX disabled

Looks like the attach API changed in Java 9 as well, so as of this patch
JMX must be enabled outside of the agent.

Tested with various host JVM versions and openjdk Docker images:
- openjdk:7-jre-alpine
- openjdk:8-jre-alpine
- openjdk:8-jre-slim (there's no alpine build yet)

Java 9 in agent will need a separate fix.