Detect invalid values in cpuacct.usage. SMAGENT-761
One possible situation is resetting the usage by manually writing
0 to cpuacct.usage, another (not fully debugged but it's the only
explanation) is apparent temporary reads of zero that then
return to the real value with the next read.

To properly track these, we need to remember at least two older
samples. In general we return the value (current - last), but when

1. last is lower than both the current value and the one before
   the last ("prev")

and

2. the current value is higher than prev

i.e. cpu usage is growing monotonically except for a single-sample
dip, the unexpected lower value is ignored and we return current - prev.

Also, log a warning when the cpuacct value drops, even without a dip
(i.e. the counter has been reset).