Don't log output for successful build steps (#717)
Travis has a 4MB limit on log length, and we are now hitting that length
regularly. To reduce the log size, wrap the two longest
steps (build/install) in a script that first writes the output to a
file, and prints the log only if the build step fails.

The invocation of the script is in turn invoked using "travis_wait",
which emits periodic log lines so travis doesn't time out the build due
to inactivity.