Handle missed samples when reading CPU usage. SMAGENT-774
If the cpuacct scan does not happen exactly every 1 second,
we misreport the CPU usage which leads to spikes.

Divide the reported delta by the delta in timestamps.
Base the time on jiffies, to match `docker stats` and scale
the result of sinsp_procfs_parser::read_cgroup_used_cpuacct_cpu_time
to be in percent (of a single cpu, so may be above 100%).
Further scaling (100->10000) happens in sinsp_analyzer::emit_container.

Clamp the reported values to 100 * n_cpus.

Also, log cpuacct scans more than 1.1 sec apart.  This should probably
be handled elsewhere, on a level as high as possible, to avoid
duplicating this logic in every place we rely on doing things every
second.