Preserve order of shm states when reader is active (#728)
When switching states, the memdumper would check to see if a capture job
was reading the first state and if so not recycle it to the back. This
wasn't conservative enough--it's possible that the capture job iterator
might go through the set of states more than once.

To fix this, check to see if a capture job is reading through the set of
states at all, and if so go through the steps of adding a third state or
entirely skipping event concatenation. That way, the reader is
guaranteed to not loop.

This fixes SSPROD-496.