Add is different addr tests (#744)
* Better code sharing for udp client/server tests

There was a lot of duplicated steps for setting up the udp server and
client with various options. Add a wrapper class udp_servers_and_client
which holds the lists of threads/servers and handles startup/running to
send the messages.

This is especially useful for later changes where tests start up more
than one server to test sendto/recvfrom changing to different servers on
a single socket.

* Add tests for fd.name_changed()

A new test udp_client_server_fd_name_changes sets up a moderately
complicated test similar to udp_client_server_sendmsg_2buf, but also
adds a second server. The client sends messages to the first server and
then the second.

The test looks for events where fd.name_changed is true, which indicates
events where the fd's connection information changes. This occurs for
the binds, initial sends/recvs on each socket in each direction, and
cases when a sendto/recvfrom occurs for a new message on an existing
socket.

* Add additional tests

Count the number of fd.name_changed events for every test case instead
of just the sendmsg/recvmsg test. There should be 7 fd.name_changed
events except for the test using connect, which skips the name changing
at the client side as a part of the recvfrom().

Also add a test that checks multiple connect() calls on the same
fd. Generally every connect() results in a name change other than when
the same address + port is specified on an already connected socket.