Add unit tests for setpgid() event handling (#778)
* Add unit tests for setpgid() event handling

The tests do the following:

 - Start a child process. The child starts a new session (and process
   group).
 - The child spawns a grandchild. The grandchild does the following:
    - calls setpgid(0, 0) to change to its own process group
    - calls setpgid(getpid, child_pid) to change back to the child's
   process group.

There are 2 variants of this. One runs in the global namespce and one
runs a pid namespace.

We want to verify that after the two setpgid() events that the caller's
threadinfo has the right m_vpgid values. After the first, it should be
the caller's pid, and after the second, it should be the child's pid (or
1, for the variant in a pid namespace).

* Wait for execve()s to check pgid

We're eliminating setpgid() parsing in
https://github.com/draios/sysdig/pull/1080 to simplify tracking of pgid,
at the expense of the pgid only being updated after an execve().

Update the tests to handle this. After each setpgid(), do an execve()
shortly afterward. The first execve() is of test_helper with a new
argument "pgid_test". test_helper + pgid_test does a second setpgid()
and execve() of /bin/echo. (Unlike other test_helper invocations, this
is done directly and not through start_process, so we skip the steps of
writing/reading stdin to synchronize with the parent).

When analyzing events, look at execve instead of setpgid before checking
the process's vpgid.