K8s daemonset (#799)
* FEAT: Support to detect & pickup changes to configuration file(s)

- The configuration files are read/parsed only during initilization. If a
configuration file is changed, the only way to pickup the changes are to
manually restart the agent container/process.
- With these changes, the agent monitors the configured list of files and
detects when they are changed. On detecting change to a monitored file,
all the agent processes are restarted to pickup the latest configuration -
similar to what happens when dragent.auto.yaml file is received from
backend.

	modified:   userspace/dragent/configuration.cpp
	modified:   userspace/dragent/configuration.h
	modified:   userspace/dragent/dragent.cpp
	modified:   userspace/dragent/dragent.default.yaml
	modified:   userspace/dragent/dragent.h

* Add new config file: /opt/draios/etc/kubernetes/config/dragent.yaml
and /opt/draios/etc/kubernetes/secrets/access-key

They are meant to be populated by Kubernetes Configmaps and Secrets
respectively. Agent will monitor them and restart if any change happens

* Minor lifting

* If file gets deleted just restart as well rather than keep going with Warnings

* Fix config file name

* Delete daemonsets since they are available in another repo

* Update dragent config map reload interval