Only consider falco events for syscall baselines (#808)
When enforcing syscall baselines, only consider those events that falco
would consider. This allows using a single restricted set of system
calls that capture important activity, and also allows baselines to be
more compatible with falco.

Baseline fingerprints contain all syscalls as before. Instead, within
the agent, the set of syscalls is filtered to only consider those
syscalls matching falco_consider(). This is done both when populating
the set of hash tables used for event/syscall id matching and when
attempting to match a sinsp_event against the set of syscall policies.