Secure perf improvements (#837)
* Use get_container based on pointers, not objects

Modify uses of get_container to return pointers to
sinsp_container_info objects instead of creating one and filling it
in.

This depends on corresponding sysdig changes.

* Make qualifies filter slightly more efficient

Make the qualifies filter slightly more efficient by doing the proc
comparison first. A better fix is in
https://github.com/draios/agent/pull/832, which gets rid of the filter
entirely and relies on a method instead. This will be more efficient
once get_container_info returns a pointer as it will in a
corresponding set of sysdig changes.