Add sanity check for tinfo->get_main_thread()
The sanity check did not fire even once but it seems to prevent
(some) crashes, possibly due to reduced number of get_main_thread()
calls, even though we're single-threaded here and should not race
with anything.

This does *not* fix the bug (suspected memory corruption)
but papers over it enough to make a difference.