Scap process blacklist (#861)
* Make suppressed comms configurable

New config option skip_events_by_process contains a list of comm
values. Those are passed to the inspector in sinsp_worker after creating
the inspector and before opening the source.

* Unit tests for suppressed threads

Testcases include:

 - Starting a program that should be suppressed by name before the
   inspector is started
 - Starting a program that should be suppressed by name after the
   inspector is started
 - Adding a comm to the set of suppressed comms after the inspector has
   started (and before the process has started)
 - Starting a chain of processes/subprocess and ensuring that the whole
   chain is suppressed via parent/child associations.

The validations ensure that a quotactl done by a suppressed program does
not appear and that a procexit event is always observed with the tid not
in the set of suppressed tids.

Additionally, at the end of the test it checks that some events were
suppressed and that no threads remain in the set of suppressed tids.

* Add addl hooks to event capture framework

Modify the event capture framework to add a function that can access the
inspector after it's been created but before it opens anything.

Add a hook before_close that is run after the capture is complete but
before the inspector is deleted.

Make the number of SCAP_TIMEOUTS before exiting the inspector loop
configurable. Necessary when suppressed threads/processes lead to a lot
of SCAP_TIMEOUTS.

This is combined with a new hook capture_continue that can be used to
stop the capture loop instead of relying on timeouts.

* Add ability to run test_helper tests in threads

If the first argument to test_helper is "threaded", the actions
specified by the following arguments will be performed in a thread
instead of directly by the test_helper process. Useful for some tests
where you want to observe clones and multiple tids.

* Add a shell script wrapper for test_helper

Add wrapper script around test_helper. Useful for when you want
test_helper spawned by another program.

* Log the number of suppressed tids

In the event counts log line, also log the number of suppressed
tids.

Also log the number of suppressed threads in sinsp_worker's log
messages. This required several changes as it's an analyzer callback
that presents these counts.

Will be used to observe when/if a set of suppressed threads grows
unexpectedly due to parent-child relationships.

* Fix valgrind errors

m_mounts_limit_size wasn't initialized. This doesn't come up in the full
agent as it's set from the dragent-level configuration.