Generalize coclient for arbitrary rpcs (#893)
* Generalize coclient for arbitrary rpcs

The current coclient interface requires copying nearly-identical code
for every new use case, which is wasteful. In addition, we want to try
to separate the connection to the server from uses of the connection,
to allow for a single connection with multiple users.

Avoid this by coming up with a templated version, with the following
classes:
 - grpc_client_defaults
 - grpc_connect, a function to create a grpc stub
 - unary_grpc_client_ + a define to create one
 - streaming_grpc_client_ + a define to create one

First, a generic grpc_client_defaults is not templated and holds the
default socket location and number of messages to read in a call to
process_queue().

grpc_connect is templated by the grpc stub name and returns a shared_ptr
of the stub returned by grpc::CreateChannel.

unary_grpc_client_ templates the request message, the response message, the
reader object that reads responses, and the interface to create a
stub. Its main methods are:
  - do_rpc(): takes the request message and does a
    non-streaming rpc of that method. The response is returned
    via a call to the callback function response_cb.
  - process_queue(): call next() to read results.
  - next(): reads from the completion queue to determine if the call
    succeeded or failed, reads the response into the appropriate
    proto object, and calls response_cb().

In both cases, the callback response_cb returns a bool "successful" that
tells you whether the rpc completed (for unary cases) or whether the
server had an error instead of returning the next response
message (streaming cases). You have the option of providing the callback
function either at rpc call time or when calling process_queue().

In do_rpc(), immediately after calling the method, call process_queue(), which in
turn calls m_cq->AsyncNext(), which moves along the process of sending
the request and possibly reading the response. This reduces the chances
of a delay/timeout if the user of streaming_grpc_client isn't calling
process_queue() very often.

To avoid the cumbersome template arguments, there's a convienence define
unary_grpc_client(Method) which simply takes the method you want to call
and uses decltype() to deduce the types of the request, response, and
stub from the method. This is roughly based on
https://stackoverflow.com/questions/15148749/pointer-to-class-member-as-a-template-parameter
which suggests that C++17 has a cleaner solution.

streaming_grpc_client_ and associated define
streaming_grpc_client(Method) are the same, but work for a streaming rpc
call instead of a unary rpc call. They are different classes as the
underlying grpc calls are slightly different with respect to state
management and how the results are read.

The streaming variant also has a state machine to handle the phases of
initial state, sending the rpc from the client, reading responses from
the server, etc.

The streaming variant additionally has a method stop_rpc() that cleanly
shuts down the client via TryCancel(). This is done non-blocking, so it
will result in an error when reading from the completion queue that
indicates that the connection has been closed. To differentiate from
true errors, use a state STATE_CLOSING. In this state, when
AsyncNext returns with updates_ok==false, instead of logging an error
simply change the state to STATE_CLOSED.

When destroying the client, call stop_rpc(). This is important, as grpc
holds state related to in-progress calls e.g. connection information and
on process exit it can take up to 10 seconds waiting for the teardown to
time out.

Note that there is no variant for bidirectional streaming, as we don't
have a need for it yet.

After merging this change, we should clean up the agent to migrate away
from the monolithic coclient() towards these method-specific classes.

* Get rid of grpc_client_defaults

The default socket wasn't used anywhere anyway, and a cap on the number
of events to fetch per call to process_queue can easily be done as a
default argument to the method.

* Remove m_print

Not used in templated classes.

* Use state enum in streaming cb vs successf bool

Instead of returning a simple successful bool in the callback for
streaming APIs, return a state enum (OK, ERROR, SHUTDOWN). It's ERROR if
the connection was terminated before the rpc call completed, OK
afterward, and SHUTDOWN while reading responses.

This keeps some of the state management within the interface class while
providing visibility to the callbacks.