Fix container prefix test (#921)
* Switch from alpine to cirros for tests

Recently, the image id for alpine/3.5 changed, which is a bit
unexpected, as the version was unchanged.

To help mitigate this, use https://hub.docker.com/r/library/cirros/
which is another semi-standard "small" image. Hopefully its image id
will remain unchanged for a specific version.

* Set dropping mode == 1

Set drop mode to 1 in the test, which excludes a big class of events
that are not necessary for the test. This helps reduce the chances of
dropped events in the face of fetching container information, which is
currently blocking.