Fix dumper thread races (#907)
* Open sinsp_dumper in main inspector thread

When starting a capture job that does *not* have a back-in-time
component, open the dumper in the main inspector thread instead of the
capture job handler thread. This ensures that when the dumper reads the
thread list as a part of open it doesn't race with the main inspector,
who might be creating/destroying threads as it processes events.

This fixes https://sysdig.atlassian.net/browse/SMAGENT-949.

* Add tests to provoke capture job thread races

Add a variant of max_outstanding_dumps that requests captures without
any back-in-time component. This ensures that if there are thread races
while initializing the capture, they are more visible.

* Skip proc scan when opening dumper

When creating a sinsp_dumper as a part of a capture job request, read
threads from sinsp instead of from /proc.

This significantly speeds up the time it takes to start a capture job.

This depends on https://github.com/draios/sysdig/pull/1164.

* Start capture jobs through sinsp worker

Previously, capture jobs were started in the context of the thread
running the tests and not the sinsp_worker thread (which uses the
inspector). This didn't match what happens in the agent, where capture
job requests were either triggered directly from the sinsp_worker
thread or from the connection_manager via a queue. And as
https://sysdig.atlassian.net/browse/SMAGENT-949 points out, it's
important to open the sinsp_dumper in the same thread that is using
the inspector to read events.

With that in mind, update the test to schedule job requests in the
test_sinsp_worker thread. This requires adding queues to pass job
requests to test_sinsp_worker and receive responses. This is done via
a pair of queues m_dump_job_{requests,responses}. A periodic task
m_job_requests_interval reads from the job requests queue, calls
capture_job_handler->queue_job_request, and returns the response.