Load system default SSL CA certificates when available
The agent has relied on a single, self-signed sysdigcloud CA certificate
previously, but that causes problems when we use the static collector
with a public-CA certificate. Use OPENSSLDIR/certs if we can find it.

Manipulating individual certificates required upgrading to poco 1.9.0

Added "ca_cert_dir: /path/to/certs" config so the user can override