Slim agent image (#918)
* Generate dockerfiles via cmake + templating

Instead of having nearly identical copies of
Dockerfile/docker-entrypoint.sh checked in for the three variants
stable/dev/local, have a single templated version with options that
generate the right version.

A new build target jinja_venv creates a standalone python venv that
pip installs jinja2. It's used by generate_agent_docker.py that creates
a Dockerfile using the template and arguments based on the variant:

 - base_docker_image: the docker image used in the FROM command
 - sysdig_repository: the build variant to install in the
   image (e.g. stable vs dev)
 - include_agent_package: whether to use apt-get to install the agent
   package, copy one locally, or not add any agent at all.
 - build_kernel_module: if false, the Dockerfile will not install any
   packages related to building the kernel module and the entrypoint
   will not call sysdigcloud-probe-builder to install one. It is passed
   through to the entrypoint as SYSDIG_BUILD_KERNEL_MODULE, and controls
   whether or not to call the probe loader script.
 - jdk_debian_release: the debian channel from which to get the jdk.
 - launch_dragent: is passed through to the entrypoint as
   SYSDIG_LAUNCH_DRAGENT, and controls whether to launch the agent or
   simply exit the entrypoint script.
 - agent_version: the version string to use when creating the kernel
   module. If not specified, will run dragent --version to find
   the version.

When building the kernel module and not installing the agent package,
add the dkms package. This is a dependency of the agent package, but
is needed by the probe builder script.

Some options are forward-thinking for new image variants, which will be
in the next commit.

docker-entrypoint.sh now conditionally sets up /usr/src links and loads
the kernel module based on the presence of
SYSDIG_BUILD_KERNEL_MODULE. There's one copy in the docker directory,
and it's copied to the specific docker image directories at build time.

* New smaller agent image variants w/o dkms

Add a new agent image sysdig/agent-slim that is smaller and as a result
less likely to have exposed vulnerabilities. Notable changes:

 - Use bitnami/minideb:jessie instead of debian:unstable. This switches
   the debian release from unstable to jessie, which has quicker
   security updates. Additionally, bitnami/minideb is a smaller version
   of the main debian images that have unnecessary packages removed. See
   https://github.com/bitnami/minideb for more details.

 - Don't include anything that can build the kernel module (gcc, dkms,
   supporting packages).

Also add a new agent image sysdig/agent-kmodule that *only* builds
and loads the sysdigcloud_probe kernel module and does not start the
agent. The agent-kmodule variant needs to run the probe builder but doesn't
have a dragent executable to return the version. In this case, directly
specify the agent version from cmake through the templating script as
--agent-version, which is turn is set as an environment variable handled
by the probe builder script.

This depends on https://github.com/draios/sysdig/pull/1166.

The intent is that customers who are sensitive to vulnerabilities will
use the slim and agent-kmodule images. Agent-kmodule is short lived so any
vulnerabilities it creates are short lived.

* Get working with docker builder

Run python from the dependencies directory, as that's the known python
version.

Copy the generated Dockerfile/docker-entrypoint.sh files out of the
build area by adding install() steps. They are in their own component
"dockerfiles", and end up in /opt/draios/docker when installed.

Also, when generating the container using the container argument to the
entrypoint, copy the generated Dockerfiles/docker-entrypoint.sh from the
build area. They don't exist in the source directory any longer.

In release-agent, instead of using docker hub build hooks, which require the
Dockerfile/entrypoint to be in the source code, build the images locally
and push them to docker hub.

The dockerfiles are in a tarball created during the previous "docker run
agent-builder ... release" step. Untar that tarball to get the generated
dockerfiles, go into each directory, and run "docker run -t <tag> .".

Each image, including the agent image if running the stable variant, is
also tagged with a ":latest" tag.

* Get working with jenkins release-agent-rc

Now that building the rc release has been separated from releasing the
rc agent, we need to change how the docker images are built.

upload-agent, which pushes a single agent variant (always rc in our use,
but it could also be variants like stable/dev), also does a docker hub
build of that variant by taking the built Dockerfiles, changing to the
right directory, and doing a docker build and push.

release-agent-rc pushes the agent variant (stable/dev, after the
rpms/debian packages have been copied in place), as well as the
agent-kmodule/slim docker images, which depend on the stable debian
packages.

* Only build local container w/ container argument

There's a bit of overlap between the release, package, and container
targets. As one step towards removing this overlap, only build the local
docker image in the container target.