Fixes for custom container logic detection. SMAGENT-954 (#920)
* Use the matched process for environment variables. SMAGENT-954

When we walk up the process tree to find one with the matching environment,
we used to ignore the actual matching process (which could be an ancestor
of the process being tested). Now we pick up the environment of the actual
matched process to construct the container metadata.

* Keep collecting custom container metadata even after first process. SMAGENT-954

Even if we already have a custom container, keep trying to render
the name/image/labels from all processes in the container, until:
- we get all the metadata
- 10 seconds (currently hardcoded) run out

* More custom container tests. SMAGENT-954

Add tests for large environments and for collecting the container
metadata across multiple processes.

* Make incremental metadata configurable due to potential perf hit. SMAGENT-954

Note: we still continue the metadata scan when the current process did not
set the container name.