Make dev/stable/rc truly orthogonal from image (#952)
Originally, the agent-slim/agent-kmodule images were only going to be
for the stable repo. Late in the changes for
https://github.com/draios/agent/pull/918 I started to make them
orthoginal, so you could have -rc and -dev versions of the
agent-slim/agent-kmodule images, but this wasn't properly thought out
and it didn't work once we created actual builds.

Fix this so you can have a dev, stable, or rc variant for any of agent,
agent-slim, and agent-kmodule. This involves slighly reducing the number
of directories below agent/docker (agent, agent-slim, agent-kmodule),
but having a foreach that creates multiple binary directories/targets
for each of dev, stable, rc for the given image.

generate_agent_docker.py now takes two important arguments, namely the
image and the repo, setting the right template params for each.

release-agent-rc/upload-agent now loop over each image agent,
agent-slim, agent-kmodule, and use the provided REPOSITORY_BASENAME to
construct the correct tag.

docker/local is still a special case that assumes dev and only has an
agent image and not -kmodule or -slim.