Ensure that nodes and deployments send update events when their metrics change (#956)
The comparison function was bugged, and we wouldn't report some updated events
for metric changes. This makes sure that we check all fields that we use for
reporting metrics. Also, it adds the first UTs for cointerface. We call
"go test" during the build process currently, so the tests should be true
short-running UTs.

This is a first step. We need to check the other components, and we should
refactor this so we don't have so much duplicated code for each component.