Report incomplete connections to collector. SMAGENT-1028
Since now we split the connections into active and incomplete,
either of these lists may be shorter than the limit (or even
completely empty).

Note: this replicates the earlier logic of considering
TOP_CONNECTIONS_IN_SAMPLE for each of the criteria (by number
of subconnections, by total bytes and now the incomplete
connections by number of subconnections). This means that we've
gone from maximum of 2 * TOP_CONNECTIONS_PER_SAMPLE to 3 times
the value