Fix cmake dependencies for files used in building the agent image
CMake targets don't propagate file-level dependencies, so we have to add
generate_agent_docker.py and Dockerfile.jinja2 as explicit dependencies
so that changing those files will trigger rebuilding the Dockerfiles.

From add_custom_command() docs
If DEPENDS specifies any target (created by an ADD_* command) a target-level
dependency is created to make sure the target is built before any target using
this custom command. Additionally, if the target is an executable or library a
file-level dependency is created to cause the custom command to re-run whenever
the target is recompiled.