Immediately push :dev tags when doing dev release (#965)
In addition to the version tag e.g. 0.99.4049dev, we want to also create
a tag :dev for each of agent, agent-slim, agent-kmodule. This is similar
to what we do in release-agent-rc, but this occurs at build time, not
rc-to-release time, and it's for the dev tag, not latest.

So if the repository is dev, immediately push the version to the tag
"dev".