Add docker binary to agent container. (#969)
* Add docker binary to agent container.

The docker-bench compliance module needs a docker binary, so install
docker.io to get one.

It's in backports like the jdk, so change the config option from
jdk_debian_release to backports_debian_release.

* Honor SYSDIG_HOST_ROOT for docker socket

When running in a container, the docker socket is at an alternate
location. That alternate location needs to be passed to the docker-bench
command and the docker executable by setting the environment variable
DOCKER_HOST.

* Add iproute to get ss command

docker-bench needs ss, which is available in iproute2.